removed integer parameter

+++ db/schema.sql
@@ -4,10 +4,8 @@ USE burger_db;
 
 
 
-
-
 CREATE TABLE burger (
-  id INTEGER(11) AUTO_INCREMENT NOT NULL,
+  id INTEGER AUTO_INCREMENT NOT NULL,
   burger_name VARCHAR(30) NOT NULL,
   devoured BOOLEAN NOT NULL,
   PRIMARY KEY (id)
